[ProcTHOR] Small bugfix in InsideOf

+++ semantic_digital_twin/src/semantic_digital_twin/reasoning/predicates.py
@@ -428,7 +428,7 @@ class InsideOf(SpatialRelation):
         inside = mesh_b_bbox.contains(mesh_a.vertices)
         if len(inside) == 0:
             return 0.0
-        return sum(inside) / len(inside)
+        return float(sum(inside) / len(inside))
 
 
 @dataclass
